[partial flush][polishing API and documentation]
